refactor(components): move toolbar to root of custom
fix #284
