--- packages/components/src/custom/SearchBar/index.tsx
@@ -1,3 +1,0 @@
-import StyledSearchBar from './SearchBar';
-
-export { StyledSearchBar };
