+++ packages/components/src/custom/StyledSearchBar/StyledSearchBar.tsx
