+++ packages/components/src/custom/StyledSearchBar/index.tsx
@@ -1,0 +1,3 @@
+import StyledSearchBar from './StyledSearchBar';
+
+export { StyledSearchBar };
