--- packages/components/src/custom/Toolbar/index.tsx
@@ -1,5 +1,0 @@
-import CustomColumnVisibilityControl from './custom-column';
-import UniversalFilter from './custom-filter';
-import SearchBar from './custom-search';
-
-export { CustomColumnVisibilityControl, SearchBar, UniversalFilter };
