+++ packages/components/src/custom/index.tsx
@@ -2,21 +2,15 @@
 import { StyledDialog, StyledDialogActions, StyledDialogContent } from './Dialog';
 import { useWindowDimensions } from './Helpers/Dimension';
 import { useNotificationHandler } from './Helpers/Notification';
-import CustomColumnVisibilityControl from './Toolbar/custom-column';
-import UniversalFilter from './Toolbar/custom-filter';
-import SearchBar from './Toolbar/custom-search';
 import { StyledTooltip } from './Tooltip';
 
 export { StyledChartDialog } from './ChartDialog';
-export { StyledSearchBar } from './SearchBar';
+export { StyledSearchBar } from './StyledSearchBar';
 export {
-  CustomColumnVisibilityControl,
-  SearchBar,
   StyledDialog,
   StyledDialogActions,
   StyledDialogContent,
   StyledTooltip,
-  UniversalFilter,
   useNotificationHandler,
   useWindowDimensions
 };
